Add --parallel option
